+++ tests/test_regex_matcher.py
@@ -611,13 +611,14 @@ def test_set_find_progress_callback() -> None:
     #       int64_t matchIndex
     # )
     def _find_progress_callback1(_context: object, _match_index: int) -> bool:
-        assert _context is None
+        _ = _context
         nonlocal result1
         result1.append(_match_index)
         return True
 
     def _find_progress_callback2(_context: object, _match_index: int) -> bool:
-        assert isinstance(_context, list)
+        if not isinstance(_context, list):
+            return False
         _context.append(_match_index)
         return _match_index < 5
 
@@ -631,15 +632,17 @@ def test_set_find_progress_callback() -> None:
     #       UErrorCode &status
     # )
     callback0, context0 = matcher.get_find_progress_callback()
-    assert isinstance(callback0, icu.URegexFindProgressCallbackPtr)
+    assert isinstance(callback0, icu.URegexFindProgressCallback)
+    assert bool(callback0) is False
     assert isinstance(context0, icu.ConstVoidPtr)
+    assert context0.value() is None
 
     # void icu::RegexMatcher::setFindProgressCallback(
     #       URegexFindProgressCallback *callback,
     #       const void *context,
     #       UErrorCode &status
     # )
-    callback1 = icu.URegexFindProgressCallbackPtr(_find_progress_callback1)
+    callback1 = icu.URegexFindProgressCallback(_find_progress_callback1)
     context1 = icu.ConstVoidPtr(None)
     matcher.set_find_progress_callback(callback1, context1)
     assert matcher.find() is True
@@ -648,11 +651,13 @@ def test_set_find_progress_callback() -> None:
     assert result1 == [1, 5, 6]
 
     callback1a, context1a = matcher.get_find_progress_callback()
-    assert isinstance(callback1a, icu.URegexFindProgressCallbackPtr)
+    assert isinstance(callback1a, icu.URegexFindProgressCallback)
+    assert bool(callback1a) is True
     assert isinstance(context1a, icu.ConstVoidPtr)
+    assert context1a.value() is None
 
     result2: list[int] = []
-    callback2 = icu.URegexFindProgressCallbackPtr(_find_progress_callback2)
+    callback2 = icu.URegexFindProgressCallback(_find_progress_callback2)
     context2 = icu.ConstVoidPtr(result2)
     matcher.set_find_progress_callback(callback2, context2)
     assert matcher.find(0) is True
@@ -660,6 +665,12 @@ def test_set_find_progress_callback() -> None:
         matcher.find()
     assert exc_info.value.args[0] == icu.UErrorCode.U_REGEX_STOPPED_BY_CALLER
     assert result2 == [1, 5]
+
+    callback2a, context2a = matcher.get_find_progress_callback()
+    assert isinstance(callback2a, icu.URegexFindProgressCallback)
+    assert bool(callback2a) is True
+    assert isinstance(context2a, icu.ConstVoidPtr)
+    assert context2a.value() == result2
 
     result1.clear()
     result2.clear()
@@ -677,6 +688,25 @@ def test_set_find_progress_callback() -> None:
     assert matcher.find() is False
     assert result1 == [1, 5, 6]
     assert len(result2) == 0
+
+    result1.clear()
+    result2.clear()
+    matcher.set_find_progress_callback(callback2a, context2a)
+    assert matcher.reset()
+    assert matcher.find() is True
+    with pytest.raises(icu.ICUError) as exc_info:
+        matcher.find()
+    assert exc_info.value.error_code == icu.U_REGEX_STOPPED_BY_CALLER
+    assert result2 == [1, 5]
+
+    result1.clear()
+    result2.clear()
+    matcher.set_find_progress_callback(callback2, context0)
+    assert matcher.reset()
+    with pytest.raises(icu.ICUError) as exc_info:
+        matcher.find()
+    assert exc_info.value.error_code == icu.U_REGEX_STOPPED_BY_CALLER
+    assert len(result1) == len(result2) == 0
 
 
 def test_set_match_callback() -> None: